♻️refactor: image 클래스명 TravelPhoto 으로 변경
